requirements.txt
numpy>=1.19.0
scipy>=1.7.0
filterpy>=1.4.5
opencv-python>=4.5.0
pyyaml>=5.4.0
pytest>=6.0.0
torch>=1.7.0
torchvision>=0.8.0
ultralytics>=8.0.0
gdown>=4.7.1